necesito una mujer estoy todo el dia programando

+++ cPedido.cs
@@ -13,7 +13,7 @@ namespace tp_final
         public float largo { get; set; }
         public float ancho { get; set; }
         public float alto { get; set; }
-        public int volumen { get; set; }
+        public float volumen { get; set; }
         public int peso { get; set; }
         public int prioridad { get; set; }
         public string? barrio { get; set; }
